requirements.txt
Pillow==9.3.0
schedule==1.1.0
pyserial==3.5
